expose DB2BulkCopy.ProviderSpecificCopyImpl
