+++ Source/LinqToDB/DataProvider/Informix/InformixBulkCopy.cs
@@ -40,7 +40,7 @@ namespace LinqToDB.DataProvider.Informix
 							connection,
 							_provider.Adapter.InformixBulkCopy);
 					else
-						return DB2.DB2BulkCopy.ProviderSpecificCopyImpl(
+						return DB2.DB2BulkCopyShared.ProviderSpecificCopyImpl(
 							table,
 							options.BulkCopyOptions,
 							source,
@@ -74,7 +74,7 @@ namespace LinqToDB.DataProvider.Informix
 							connection,
 							_provider.Adapter.InformixBulkCopy));
 					else
-						return Task.FromResult(DB2.DB2BulkCopy.ProviderSpecificCopyImpl(
+						return Task.FromResult(DB2.DB2BulkCopyShared.ProviderSpecificCopyImpl(
 							table,
 							options.BulkCopyOptions,
 							source,
@@ -113,7 +113,7 @@ namespace LinqToDB.DataProvider.Informix
 								connection,
 								_provider.Adapter.InformixBulkCopy);
 						else
-							return DB2.DB2BulkCopy.ProviderSpecificCopyImpl(
+							return DB2.DB2BulkCopyShared.ProviderSpecificCopyImpl(
 								table,
 								options.BulkCopyOptions,
 								syncSource,
